+++ src/main/java/tvgirl/elmodev/e1m0Admin/api/service/gui/SecretCodeServiceAPI.java
@@ -5,7 +5,7 @@ import tvgirl.elmodev.e1m0Admin.state.secretcode.SecretCodeState;
 import java.util.UUID;
 
 public interface SecretCodeServiceAPI {
-    void oneStepHandler(UUID id, byte i);
+    void oneStepHandler(UUID id, int i);
 
     String getInputCode(UUID id);
 }
